# mgr/dashboard: force TLS 1.3