@epuertat 2023-03-13
Failures:
* [Dashboard e2e](https://jenkins.ceph.com/job/ceph-dashboard-pull-requests/12327/):
  * Applitools: diff in heartbeat menu icon. [Fixed](https://eyes.applitools.com/app/test-results/00000251723622279366/?accountId=hCTedwnSTUWSCZlXPqua0g~~)
  * RBD mirroring:
```
1) Mirroring page
       rbd mirroring bootstrap
         should generate and import the bootstrap token between clusters:
     AssertionError: Timed out retrying after 120000ms: Expected to find element: `cd-pool-list`, but never found it.
      at Context.eval (https://172.21.5.38:7820/__cypress/tests?p=cypress/integration/block/mirroring.e2e-spec.ts:217:21)

  2) Mirroring page
       rbd mirroring bootstrap
         "after each" hook for "should generate and import the bootstrap token between clusters":
     HttpErrorResponse: The following error originated from your application code, not from Cypress.

  > Http failure response for https://172.21.5.38:7820/ui-api/standard_settings: 401 Unauthorized
```
* Cephadm e2e: [`should check if rgw daemon is running on all hosts.Host Page should check if rgw daemon is running on all hosts (from Host Page)`](https://jenkins.ceph.com/job/ceph-dashboard-cephadm-e2e/6606/testReport/junit/(root)/should%20check%20if%20rgw%20daemon%20is%20running%20on%20all%20hosts/Host_Page_should_check_if_rgw_daemon_is_running_on_all_hosts/)