@epuertat 2023-03-13
jenkins test dashboard cephadm